Refactor: Extract GradeVerificationController from ElectronicGraderController
